fixed types

+++ app/queue.go
@@ -2,7 +2,7 @@ package app
 
 import "github.com/webitel/call_center/model"
 
-func (a *App) GetQueueById(id int) (*model.Queue, *model.AppError) {
+func (a *App) GetQueueById(id int64) (*model.Queue, *model.AppError) {
 	if result := <-a.Srv.Store.Queue().GetById(id); result.Err != nil {
 		return nil, result.Err
 	} else {
